backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
motor==3.6.0
pymongo==4.9.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Quantitative Finance
numpy==1.26.4
scipy==1.14.1
pandas==2.2.3
statsmodels==0.14.4
arch==7.1.0

# Data Sources
yfinance>=0.2.58
requests==2.32.3
websockets==13.1
openbb>=4.7.0
openbb-yfinance>=1.4.0
openbb-cboe>=1.4.0

# Machine Learning
scikit-learn==1.5.2
torch==2.4.1
transformers==4.45.2
torchsde==0.2.6
einops==0.8.1
huggingface-hub==0.33.1
tqdm==4.67.1
safetensors==0.6.2

# NLP
spacy==3.8.2
textblob==0.18.0.post0

# Visualization & Reports
matplotlib==3.9.2
reportlab==4.2.4

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
aiohttp==3.10.9
celery==5.4.0
redis==5.1.1

# Rate limiting
slowapi==0.1.9

# CORS
starlette==0.38.6

# Testing
pytest==9.0.2
pytest-asyncio==1.3.0
